test-app/requirements.txt
# URL Shortener Application Dependencies

# Core async libraries
aiosqlite==0.19.0
asyncio-mqtt==0.13.0

# Security and authentication
bcrypt==4.0.1
PyJWT==2.8.0

# HTTP and web framework dependencies (if needed for future API)
fastapi==0.104.1
uvicorn==0.24.0

# Data validation and serialization
pydantic==2.5.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development dependencies
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
click==8.1.7